Moved some functions from Framework to Game
